% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTable.R
\name{addTable}
\alias{addTable}
\title{Add a new table to a base table}
\usage{
addTable(connection, firstTable, secondTableName)
}
\arguments{
\item{connection}{A connection object to the database.}

\item{firstTable}{The base table to which the new table will be added.}

\item{secondTableName}{The name of the new table to be added.}
}
\value{
A merged table consisting of the base table and the new table.
}
\description{
This function merges a new table with a base table based on a common ID column.
It first identifies the relationship between the new table and the base table, and then
identifies the ID column that they share. It also identifies any event and time columns
in the new table. If a time column is present, the new table is sorted chronologically.
If an event column is present, the event concepts are translated, the longitudinal data
is sequenced, and the table is reshaped. Finally, the new table is merged with the base
table using the common ID column.
}
