% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTables.R
\name{addTables}
\alias{addTables}
\title{Add all tables related to 'person'}
\usage{
addTables(connection)
}
\arguments{
\item{connection}{A connection object to the database.}
}
\value{
A data frame that is the result of merging the base table with all its related tables.
}
\description{
This function retrieves all tables related to the 'person' table from the database and adds them to the base table.
The base table is initially the 'person' table and is updated iteratively with each related table.
The function uses the \code{addTable} function to merge each related table with the current base table.
}
