% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_catalog.R
\name{checkConnectionDS}
\alias{checkConnectionDS}
\title{Check Database Connection}
\usage{
checkConnectionDS(resource)
}
\arguments{
\item{resource}{An object representing the database resource.}
}
\description{
This function checks if a connection to the database can be established using the provided resource.
It is designed to be called from the DataSHIELD client to verify that the database connection is active and valid.
}
