% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_catalog.R
\name{checkPrivacyControlLevelDS}
\alias{checkPrivacyControlLevelDS}
\title{Check Privacy Control Level}
\usage{
checkPrivacyControlLevelDS()
}
\value{
An error message if the privacy control level is not set to 'permissive' or 'banana'.
}
\description{
This function checks if the current privacy control level is set to either 'permissive' or 'banana'.
It is designed to be called from the DataSHIELD client to verify that the privacy control level is permissive enough to allow the
required operations used by the package.
If the privacy control level is not set to 'permissive' or 'banana', the resulting error is returned.
}
