% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanTable.R
\name{cleanTable}
\alias{cleanTable}
\title{Clean table}
\usage{
cleanTable(table, removeConceptId = TRUE, removeNA = FALSE)
}
\arguments{
\item{table}{The input table to be cleaned.}

\item{removeConceptId}{A logical value indicating whether to remove columns with concept IDs. Default is TRUE.}

\item{removeNA}{A logical value indicating whether to remove columns with NA values. Default is FALSE.}
}
\value{
A cleaned table.
}
\description{
This function cleans a given table by optionally removing columns with concept IDs and/or columns with NA values.
}
