% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_connection.R
\name{closeConnection}
\alias{closeConnection}
\title{Close Database Connection}
\usage{
closeConnection(connection, error = NULL)
}
\arguments{
\item{connection}{A DBI database connection object.}

\item{error}{An optional error message to be thrown if not NULL.}
}
\description{
Closes an active database connection. If an error parameter is provided, the function will propagate the specified error.
}
