% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOMOPCDMTableDS.R
\name{createOMOPCDMTableDS}
\alias{createOMOPCDMTableDS}
\title{Create an OMOP CDM table for DataSHIELD}
\usage{
createOMOPCDMTableDS(resource, removeConceptId = TRUE, removeNA = FALSE)
}
\arguments{
\item{resource}{An object of class 'OMOPCDMResourceClient' representing the OMOP CDM database resource.}

\item{removeConceptId}{A logical indicating whether to remove concept IDs from the table. Default is TRUE.}

\item{removeNA}{A logical indicating whether to remove NA values from the table. Default is FALSE.}
}
\value{
A table object representing the OMOP CDM table.
}
\description{
This function is called from the DataSHIELD client to the server to retrieve an OMOP CDM table.
It creates a table from the provided resource, with options to remove concept IDs and/or NA values.
}
