% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTable.R
\name{createTable}
\alias{createTable}
\title{Create OMOP CDM table}
\usage{
createTable(resource, removeConceptId = TRUE, removeNA = FALSE)
}
\arguments{
\item{resource}{The resource to be used for creating the table.}

\item{removeConceptId}{A logical value indicating whether to remove columns with concept IDs. Default is TRUE.}

\item{removeNA}{A logical value indicating whether to remove columns with NA values. Default is FALSE.}
}
\value{
A fully processed OMOP CDM table.
}
\description{
This function creates an Observational Medical Outcomes Partnership (OMOP) Common Data Model (CDM) table.
It performs several operations such as adding tables related to the 'person' table, reshaping them and resolving their ID values.
}
