% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\docType{data}
\name{.achilles_domain_analyses}
\alias{.achilles_domain_analyses}
\title{Mapping of OMOP CDM domains to their Achilles analysis IDs}
\format{
An object of class \code{list} of length 9.
}
\usage{
.achilles_domain_analyses
}
\description{
Named list where each element is an integer vector of analysis IDs
associated with the domain.
}
\keyword{internal}
