% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{.achillesAnalysisResultTable}
\alias{.achillesAnalysisResultTable}
\title{Determine result table from analysis_id}
\usage{
.achillesAnalysisResultTable(analysis_ids)
}
\description{
Determine result table from analysis_id
}
\keyword{internal}
