% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{.achillesGetHeelResults}
\alias{.achillesGetHeelResults}
\title{Get Achilles Heel warnings}
\usage{
.achillesGetHeelResults(handle)
}
\arguments{
\item{handle}{CDM handle}
}
\value{
data.frame with analysis_id, achilles_heel_warning, rule_id,
  record_count
}
\description{
Get Achilles Heel warnings
}
\keyword{internal}
