% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles_gating.R
\name{.achillesRecordGatedIds}
\alias{.achillesRecordGatedIds}
\title{Record-unit analysis IDs that require person-gating}
\usage{
.achillesRecordGatedIds()
}
\value{
Integer vector of analysis IDs.
}
\description{
Record-unit analysis IDs that require person-gating
}
\keyword{internal}
