% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{.achillesStatus}
\alias{.achillesStatus}
\title{Check Achilles availability}
\usage{
.achillesStatus(handle)
}
\arguments{
\item{handle}{CDM handle}
}
\value{
list(available, tables, n_analyses, n_heel_warnings)
}
\description{
Check Achilles availability
}
\keyword{internal}
