% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure.R
\name{.assertPrivacyLevel}
\alias{.assertPrivacyLevel}
\title{Check if privacy control level allows an operation}
\usage{
.assertPrivacyLevel(required_level = "banana")
}
\arguments{
\item{required_level}{Character; minimum level required}
}
\value{
TRUE invisibly, or stops with an error
}
\description{
Check if privacy control level allows an operation
}
\keyword{internal}
