% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.buildJoinGraph}
\alias{.buildJoinGraph}
\title{Build join graph from OHDSI FK metadata}
\usage{
.buildJoinGraph(field_level, present_tables)
}
\arguments{
\item{field_level}{Field-level metadata data.frame}

\item{present_tables}{Character vector of tables present in DB}
}
\value{
Data frame with from_table, from_column, to_table, to_column
}
\description{
Build join graph from OHDSI FK metadata
}
\keyword{internal}
