% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.closeHandle}
\alias{.closeHandle}
\title{Close a CDM handle}
\usage{
.closeHandle(handle)
}
\arguments{
\item{handle}{A CDM handle}
}
\value{
NULL, called for side effect of closing the database connection.
}
\description{
Close a CDM handle
}
\keyword{internal}
