% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{.cohortCombine}
\alias{.cohortCombine}
\title{Combine two cohorts using set operations}
\usage{
.cohortCombine(handle, op, cohort_table_a, cohort_table_b, new_name = NULL)
}
\arguments{
\item{handle}{CDM handle}

\item{op}{Character; "intersect", "union", or "setdiff"}

\item{cohort_table_a}{Character; first cohort temp table}

\item{cohort_table_b}{Character; second cohort temp table}

\item{new_name}{Character; name for result temp table}
}
\value{
Character; result temp table name
}
\description{
Combine two cohorts using set operations
}
\keyword{internal}
