% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{.cohortGetDefinition}
\alias{.cohortGetDefinition}
\title{Get a specific cohort definition}
\usage{
.cohortGetDefinition(handle, cohort_definition_id)
}
\arguments{
\item{handle}{CDM handle}

\item{cohort_definition_id}{Integer}
}
\value{
Named list with definition details
}
\description{
Get a specific cohort definition
}
\keyword{internal}
