% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{.cohortList}
\alias{.cohortList}
\title{List available cohort definitions}
\usage{
.cohortList(handle)
}
\arguments{
\item{handle}{CDM handle}
}
\value{
Data frame with cohort definitions
}
\description{
List available cohort definitions
}
\keyword{internal}
