% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.compileConceptLookup}
\alias{.compileConceptLookup}
\title{Compile SQL for concept lookup}
\usage{
.compileConceptLookup(handle, concept_ids)
}
\arguments{
\item{handle}{CDM handle}

\item{concept_ids}{Numeric vector}
}
\value{
Character; SQL returning concept_id, concept_name
}
\description{
Compile SQL for concept lookup
}
\keyword{internal}
