% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.compilePersonCount}
\alias{.compilePersonCount}
\title{Compile SQL to count distinct persons}
\usage{
.compilePersonCount(handle, from_sql)
}
\arguments{
\item{handle}{CDM handle}

\item{from_sql}{Character; the full SELECT query}
}
\value{
Character; SQL returning single count
}
\description{
Compile SQL to count distinct persons
}
\keyword{internal}
