% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.convertTypes}
\alias{.convertTypes}
\title{Convert data types in a result data frame}
\usage{
.convertTypes(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Data frame with converted types
}
\description{
Convert data types in a result data frame
}
\keyword{internal}
