% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{.createStagingDir}
\alias{.createStagingDir}
\title{Create a staging directory for a token}
\usage{
.createStagingDir(token)
}
\arguments{
\item{token}{Character; staging token}
}
\value{
Character; path to the staging directory
}
\description{
Create a staging directory for a token
}
\keyword{internal}
