% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.createTempTable}
\alias{.createTempTable}
\title{Create a temporary table in the database}
\usage{
.createTempTable(handle, name, select_sql)
}
\arguments{
\item{handle}{CDM handle}

\item{name}{Character; temp table name}

\item{select_sql}{Character; SELECT statement for contents}
}
\value{
The temp table name
}
\description{
Create a temporary table in the database
}
\keyword{internal}
