% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.databaseVersionAtLeast}
\alias{.databaseVersionAtLeast}
\title{Compare a vendor version string with a minimum version}
\usage{
.databaseVersionAtLeast(version, minimum, family)
}
\description{
Compare a vendor version string with a minimum version
}
\keyword{internal}
