% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.detectCDMInfo}
\alias{.detectCDMInfo}
\title{Detect CDM info from cdm_source table}
\usage{
.detectCDMInfo(handle, tables)
}
\arguments{
\item{handle}{CDM handle}

\item{tables}{Character vector of available tables}
}
\value{
Named list or NULL
}
\description{
Detect CDM info from cdm_source table
}
\keyword{internal}
