% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.dropTempTable}
\alias{.dropTempTable}
\title{Drop a temporary table}
\usage{
.dropTempTable(handle, name)
}
\arguments{
\item{handle}{CDM handle}

\item{name}{Character; temp table name}
}
\value{
NULL, called for side effect.
}
\description{
Drop a temporary table
}
\keyword{internal}
