% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.executeQuery}
\alias{.executeQuery}
\title{Execute a SQL query and return a data frame}
\usage{
.executeQuery(handle, sql)
}
\arguments{
\item{handle}{CDM handle}

\item{sql}{Character; SQL to execute}
}
\value{
Data frame
}
\description{
Execute a SQL query and return a data frame
}
\keyword{internal}
