% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{.filterIndexEventEpisodes}
\alias{.filterIndexEventEpisodes}
\title{Apply population eligibility predicates to each concrete index episode}
\usage{
.filterIndexEventEpisodes(handle, bp, cohort_table, filter_tree, name)
}
\description{
Apply population eligibility predicates to each concrete index episode
}
\keyword{internal}
