% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{.generateStagingToken}
\alias{.generateStagingToken}
\title{Generate a unique staging token}
\usage{
.generateStagingToken()
}
\value{
Character; staging token
}
\description{
Generate a unique staging token
}
\keyword{internal}
