% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.generateTimeWindows}
\alias{.generateTimeWindows}
\title{Generate time window bins}
\usage{
.generateTimeWindows(bin_width, window_start, window_end)
}
\arguments{
\item{bin_width}{Integer; bin width in days}

\item{window_start}{Integer; start of window (days from index)}

\item{window_end}{Integer; end of window (days from index)}
}
\value{
Data frame with timeId, startDay, endDay
}
\description{
Generate time window bins
}
\keyword{internal}
