% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.getCapabilities}
\alias{.getCapabilities}
\title{Get capabilities signature for client validation}
\usage{
.getCapabilities(handle)
}
\arguments{
\item{handle}{CDM handle}
}
\value{
Named list with schema summary
}
\description{
Get capabilities signature for client validation
}
\keyword{internal}
