% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.getDateColumn}
\alias{.getDateColumn}
\title{Get the primary date column for a table}
\usage{
.getDateColumn(blueprint, table)
}
\arguments{
\item{blueprint}{The schema blueprint}

\item{table}{Character; table name}
}
\value{
Character; date column name, or NULL
}
\description{
Get the primary date column for a table
}
\keyword{internal}
