% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.listColumnsRaw}
\alias{.listColumnsRaw}
\title{List columns in a table (raw DB query)}
\usage{
.listColumnsRaw(handle, table, schema = NULL)
}
\arguments{
\item{handle}{CDM handle}

\item{table}{Character; table name}

\item{schema}{Character; schema name}
}
\value{
Data frame with column_name, data_type, is_nullable
}
\description{
List columns in a table (raw DB query)
}
\keyword{internal}
