% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.listTablesRaw}
\alias{.listTablesRaw}
\title{List tables in a schema (raw DB query)}
\usage{
.listTablesRaw(handle, schema = NULL)
}
\arguments{
\item{handle}{CDM handle}

\item{schema}{Character; schema name}
}
\value{
Character vector of table names (lowercase)
}
\description{
List tables in a schema (raw DB query)
}
\keyword{internal}
