% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortdx.R
\name{.ohdsiCohortDxCohortCountEntry}
\alias{.ohdsiCohortDxCohortCountEntry}
\title{Build the cohort-count entry under a given id (cohort_count / c_cohort_counts)}
\usage{
.ohdsiCohortDxCohortCountEntry(name, adapter, meta)
}
\description{
Build the cohort-count entry under a given id (cohort_count / c_cohort_counts)
}
\keyword{internal}
