% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortdx.R
\name{.ohdsiCohortDxTemporalParams}
\alias{.ohdsiCohortDxTemporalParams}
\title{Shared param spec for the temporal covariate analyses}
\usage{
.ohdsiCohortDxTemporalParams(include_value_kind = FALSE)
}
\description{
Shared param spec for the temporal covariate analyses
}
\keyword{internal}
