% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortdx.R
\name{.ohdsiCohortDxTimeSeriesEntry}
\alias{.ohdsiCohortDxTimeSeriesEntry}
\title{Build the time-series entry under a given id}
\usage{
.ohdsiCohortDxTimeSeriesEntry(name, adapter, meta)
}
\description{
Build the time-series entry under a given id
}
\keyword{internal}
