% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_results.R
\name{.ohdsiStatus}
\alias{.ohdsiStatus}
\title{Check OHDSI result tool availability}
\usage{
.ohdsiStatus(handle)
}
\arguments{
\item{handle}{CDM handle}
}
\value{
Named list with per-tool availability
}
\description{
Check OHDSI result tool availability
}
\keyword{internal}
