% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopAnalysisParamSummary}
\alias{.omopAnalysisParamSummary}
\title{Flatten a param spec list into a compact character description}
\usage{
.omopAnalysisParamSummary(params)
}
\description{
Flatten a param spec list into a compact character description
}
\keyword{internal}
