% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopCharRiskFactorSmd}
\alias{.omopCharRiskFactorSmd}
\title{\code{dsomop:char.risk_factor_smd} entry (Characterization RiskFactor)}
\usage{
.omopCharRiskFactorSmd()
}
\description{
\code{dsomop:char.risk_factor_smd} entry (Characterization RiskFactor)
}
\keyword{internal}
