% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod.R
\name{.omopCmAttritionEntry}
\alias{.omopCmAttritionEntry}
\title{\code{dsomop:ohdsi.cohort_method.cm_attrition} entry (CohortMethod attrition)}
\usage{
.omopCmAttritionEntry()
}
\description{
\code{dsomop:ohdsi.cohort_method.cm_attrition} entry (CohortMethod attrition)
}
\keyword{internal}
