% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod.R
\name{.omopCmDiagnosticsSummaryEntry}
\alias{.omopCmDiagnosticsSummaryEntry}
\title{\code{dsomop:ohdsi.cohort_method.cm_diagnostics_summary} entry (CM diagnostics)}
\usage{
.omopCmDiagnosticsSummaryEntry()
}
\description{
\code{dsomop:ohdsi.cohort_method.cm_diagnostics_summary} entry (CM diagnostics)
}
\keyword{internal}
