% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod_effect.R
\name{.omopCmEffectEstimateEntry}
\alias{.omopCmEffectEstimateEntry}
\title{\code{dsomop:cm.effect_estimate} entry (CohortMethod fitted HR / RR + CI)}
\usage{
.omopCmEffectEstimateEntry()
}
\description{
\code{dsomop:cm.effect_estimate} entry (CohortMethod fitted HR / RR + CI)
}
\keyword{internal}
