% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cm_interaction.R
\name{.omopCmInteractionEntry}
\alias{.omopCmInteractionEntry}
\title{\code{dsomop:ohdsi.cohort_method.cm_interaction_result} entry (effect modification)}
\usage{
.omopCmInteractionEntry()
}
\description{
\code{dsomop:ohdsi.cohort_method.cm_interaction_result} entry (effect modification)
}
\keyword{internal}
