% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod.R
\name{.omopCmKaplanMeierEntry}
\alias{.omopCmKaplanMeierEntry}
\title{\code{dsomop:ohdsi.cohort_method.cm_kaplan_meier_dist} entry (CM KM curve)}
\usage{
.omopCmKaplanMeierEntry()
}
\description{
\code{dsomop:ohdsi.cohort_method.cm_kaplan_meier_dist} entry (CM KM curve)
}
\keyword{internal}
