% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod.R
\name{.omopCmMdrrCanonicalEntry}
\alias{.omopCmMdrrCanonicalEntry}
\title{\code{dsomop:cm.mdrr} canonical alias of the live cm_result substrate+MDRR entry}
\usage{
.omopCmMdrrCanonicalEntry()
}
\description{
\code{dsomop:cm.mdrr} canonical alias of the live cm_result substrate+MDRR entry
}
\keyword{internal}
