% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod.R
\name{.omopCmPropensityEntry}
\alias{.omopCmPropensityEntry}
\title{\code{dsomop:cm.propensity_distribution} entry (live in-session PS substrate)}
\usage{
.omopCmPropensityEntry()
}
\description{
\code{dsomop:cm.propensity_distribution} entry (live in-session PS substrate)
}
\keyword{internal}
