% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_cohortmethod.R
\name{.omopCmResultEntry}
\alias{.omopCmResultEntry}
\title{\code{dsomop:ohdsi.cohort_method.cm_result} entry (CohortMethod substrate + MDRR)}
\usage{
.omopCmResultEntry()
}
\description{
\code{dsomop:ohdsi.cohort_method.cm_result} entry (CohortMethod substrate + MDRR)
}
\keyword{internal}
