% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopDiagCohortOverlap}
\alias{.omopDiagCohortOverlap}
\title{\code{dsomop:cohortdx.cohort_overlap} entry (CohortDiagnostics)}
\usage{
.omopDiagCohortOverlap()
}
\description{
\code{dsomop:cohortdx.cohort_overlap} entry (CohortDiagnostics)
}
\keyword{internal}
