% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopDiagTimeDistribution}
\alias{.omopDiagTimeDistribution}
\title{\code{dsomop:cohortdx.time_distribution} entry (CohortDiagnostics)}
\usage{
.omopDiagTimeDistribution()
}
\description{
\code{dsomop:cohortdx.time_distribution} entry (CohortDiagnostics)
}
\keyword{internal}
