% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopDiagTimeToEvent}
\alias{.omopDiagTimeToEvent}
\title{\code{dsomop:char.time_to_event} entry (Characterization)}
\usage{
.omopDiagTimeToEvent()
}
\description{
\code{dsomop:char.time_to_event} entry (Characterization)
}
\keyword{internal}
