% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopFeTable1}
\alias{.omopFeTable1}
\title{\code{dsomop:fe.table1} entry (FeatureExtraction createDefaultTable1)}
\usage{
.omopFeTable1()
}
\description{
\code{dsomop:fe.table1} entry (FeatureExtraction createDefaultTable1)
}
\keyword{internal}
