% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{.omopPersonKeys}
\alias{.omopPersonKeys}
\title{The person/subject token column(s) present in a frame}
\usage{
.omopPersonKeys(x)
}
\arguments{
\item{x}{A data.frame.}
}
\value{
Character vector of person-key columns present (subset of
  \code{.PERSON_KEY_COLS()}).
}
\description{
The person/subject token column(s) present in a frame
}
\keyword{internal}
