% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_plp.R
\name{.omopPlpCalibration}
\alias{.omopPlpCalibration}
\title{\code{dsomop:plp.calibration} entry (PatientLevelPrediction calibration)}
\usage{
.omopPlpCalibration()
}
\description{
\code{dsomop:plp.calibration} entry (PatientLevelPrediction calibration)
}
\keyword{internal}
