% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_plp.R
\name{.omopPlpDiagnostic}
\alias{.omopPlpDiagnostic}
\title{\code{dsomop:plp.diagnostic} entry (PatientLevelPrediction diagnostic summary)}
\usage{
.omopPlpDiagnostic()
}
\description{
\code{dsomop:plp.diagnostic} entry (PatientLevelPrediction diagnostic summary)
}
\keyword{internal}
