% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_plp.R
\name{.omopPlpPerformance}
\alias{.omopPlpPerformance}
\title{\code{dsomop:plp.performance} entry (PatientLevelPrediction performance)}
\usage{
.omopPlpPerformance()
}
\description{
\code{dsomop:plp.performance} entry (PatientLevelPrediction performance)
}
\keyword{internal}
