% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_plp.R
\name{.omopPlpThreshold}
\alias{.omopPlpThreshold}
\title{\code{dsomop:plp.threshold} entry (PatientLevelPrediction threshold summary)}
\usage{
.omopPlpThreshold()
}
\description{
\code{dsomop:plp.threshold} entry (PatientLevelPrediction threshold summary)
}
\keyword{internal}
