% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ohdsi_pack_treatmentpatterns.R
\name{.omopTxPathPathwaysEntry}
\alias{.omopTxPathPathwaysEntry}
\title{Build the treatment-pathways entry (\code{dsomop:txpath.pathways})}
\usage{
.omopTxPathPathwaysEntry()
}
\description{
Build the treatment-pathways entry (\code{dsomop:txpath.pathways})
}
\keyword{internal}
