% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onDetach}
\alias{.onDetach}
\title{Package detach hook}
\usage{
.onDetach(lib)
}
\arguments{
\item{lib}{Library path.}
}
\description{
Unregisters the OMOP CDM resource resolver on package unload.
}
\keyword{internal}
