% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{.profileDomainCoverage}
\alias{.profileDomainCoverage}
\title{Get cross-table domain coverage}
\usage{
.profileDomainCoverage(handle)
}
\arguments{
\item{handle}{CDM handle}
}
\value{
Data frame with table, n_persons, schema_category
}
\description{
Get cross-table domain coverage
}
\keyword{internal}
