% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{.profileTableStats}
\alias{.profileTableStats}
\title{Get safe table-level statistics}
\usage{
.profileTableStats(handle, table, stats = c("rows", "persons"))
}
\arguments{
\item{handle}{CDM handle}

\item{table}{Character; table name}

\item{stats}{Character vector; which stats to include}
}
\value{
Named list with requested statistics
}
\description{
Get safe table-level statistics
}
\keyword{internal}
