% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_render.R
\name{.querySql}
\alias{.querySql}
\title{Render, translate, and query SQL (returns data.frame)}
\usage{
.querySql(handle, sql, ...)
}
\arguments{
\item{handle}{CDM handle}

\item{sql}{Character; OHDSI SQL with \code{@param} placeholders}

\item{...}{Named parameters for substitution}
}
\value{
Data frame
}
\description{
Render, translate, and query SQL (returns data.frame)
}
\keyword{internal}
