% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_render.R
\name{.resolve_target_dialect}
\alias{.resolve_target_dialect}
\title{Resolve a DBMS name to a target dialect}
\usage{
.resolve_target_dialect(dbms)
}
\arguments{
\item{dbms}{Character; DBMS name from resource driver}
}
\value{
Character; target dialect string
}
\description{
Resolve a DBMS name to a target dialect
}
\keyword{internal}
