% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.resolveTableSchema}
\alias{.resolveTableSchema}
\title{Resolve which DB schema a table should be queried from}
\usage{
.resolveTableSchema(handle, table, schema_category)
}
\arguments{
\item{handle}{CDM handle}

\item{table}{Table name}

\item{schema_category}{Character; CDM, Vocabulary, or Results}
}
\value{
Character; schema name or NULL
}
\description{
Resolve which DB schema a table should be queried from
}
\keyword{internal}
