% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{.snowflakeServer}
\alias{.snowflakeServer}
\title{Normalize a Snowflake server name}
\usage{
.snowflakeServer(host)
}
\arguments{
\item{host}{Account identifier or complete Snowflake host name.}
}
\value{
Complete Snowflake host name.
}
\description{
Normalize a Snowflake server name
}
\keyword{internal}
