% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{.stagingBaseDir}
\alias{.stagingBaseDir}
\title{Get the staging base directory}
\usage{
.stagingBaseDir()
}
\value{
Character; path to staging base directory
}
\description{
Get the staging base directory
}
\keyword{internal}
