% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.standardizeName}
\alias{.standardizeName}
\title{Standardize a concept name for column names}
\usage{
.standardizeName(name)
}
\arguments{
\item{name}{Character}
}
\value{
Character; standardized name
}
\description{
Standardize a concept name for column names
}
\keyword{internal}
