% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{.vocabLookupConcepts}
\alias{.vocabLookupConcepts}
\title{Lookup concepts by ID}
\usage{
.vocabLookupConcepts(handle, concept_ids)
}
\arguments{
\item{handle}{CDM handle}

\item{concept_ids}{Numeric vector of concept IDs}
}
\value{
Data frame with concept details
}
\description{
Lookup concepts by ID
}
\keyword{internal}
