% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{.vocabTranslateColumns}
\alias{.vocabTranslateColumns}
\title{Translate concept_id columns in a data frame to concept names}
\usage{
.vocabTranslateColumns(handle, df)
}
\arguments{
\item{handle}{CDM handle}

\item{df}{Data frame}
}
\value{
Data frame with concept IDs replaced by names
}
\description{
Translate concept_id columns in a data frame to concept names
}
\keyword{internal}
