% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{.writeStagingManifest}
\alias{.writeStagingManifest}
\title{Write a staging manifest}
\usage{
.writeStagingManifest(staging_dir, descriptors)
}
\arguments{
\item{staging_dir}{Character; path to staging directory}

\item{descriptors}{Named list of descriptors}
}
\description{
Write a staging manifest
}
\keyword{internal}
