% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendColumn.R
\name{extendColumn}
\alias{extendColumn}
\title{Extend column in a table using a dictionary}
\usage{
extendColumn(table, dictionary, tableColumn, dictionaryColumn)
}
\arguments{
\item{table}{A data frame representing the table to be extended.}

\item{dictionary}{A data frame representing the dictionary used for extension.}

\item{tableColumn}{A character string representing the column in the table to be extended.}

\item{dictionaryColumn}{A character string representing the column in the dictionary used for extension.}
}
\value{
A data frame representing the table after extension.
}
\description{
This function extends a column in a table using a dictionary. It replaces the column names in the dictionary with the corresponding column in the table.
It then merges the table with the dictionary based on the common column. The function is used when extending non-sensitive values present in a table, not patient data.
Therefore, it does not require the use of the mergeDS function from dsBase.
}
