% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendTable.R
\name{extendTable}
\alias{extendTable}
\title{Extend a table with additional columns}
\usage{
extendTable(connection, table, maxDepth = 5)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{table}{The table to be extended.}

\item{maxDepth}{The maximum depth to which the table should be extended. Default is 5.}
}
\value{
The extended table.
}
\description{
This function extends a given table by adding additional columns. The columns are added based on the
matching of column names with the names in a dictionary of simplified concepts. The process continues
until either there are no more matching column names or a specified maximum depth is reached.
}
\details{
The function starts by retrieving a dictionary of simplified concepts and a list of non-related table ID
column names. It then enters a loop where it iterates over the column names. For each column name, it
checks if it matches any name in the dictionary. If a match is found, the corresponding column from the
dictionary is added to the table. The process continues until either there are no more matching column
names or the maximum depth is reached.
}
