% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_columns.R
\name{filterColumns}
\alias{filterColumns}
\title{Filter Columns from a Table}
\usage{
filterColumns(table, tableColumns, selectedColumns, keepColumns)
}
\arguments{
\item{table}{The table from which columns are to be filtered.}

\item{tableColumns}{A character vector of all column names in the table.}

\item{selectedColumns}{A character vector specifying which columns to include in the filter.}

\item{keepColumns}{A character vector of columns that must be included in the output regardless of the filter.}
}
\value{
A character vector of the filtered column names.
}
\description{
This function filters columns from a given table based on the specified criteria. It allows for the inclusion
of mandatory columns (keepColumns) regardless of the filter applied. If no specific columns are selected,
it defaults to excluding columns that contain '_source' in their names.
}
