% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClassifiedIdColumnNames.R
\name{getClassifiedIdColumnNames}
\alias{getClassifiedIdColumnNames}
\title{Classify ID column names}
\usage{
getClassifiedIdColumnNames(connection)
}
\arguments{
\item{connection}{A DBI connection object to the database.}
}
\value{
A list of character vectors. Each element of the list corresponds to a category and contains the ID column names of that category.
}
\description{
This function classifies the ID column names from a database connection based on their categories.
It retrieves the ID categories and foreign ID column names, then matches each ID column name to its category.
The result is a list where each element corresponds to a category and contains the ID column names belonging to that category.
}
