% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnNames.R
\name{getColumnNames}
\alias{getColumnNames}
\title{Get column names from database tables}
\usage{
getColumnNames(connection)
}
\arguments{
\item{connection}{A DBI connection object to the database.}
}
\value{
A list of character vectors. Each element of the list corresponds to a table in the database,
and contains the column names of that table, each prefixed with the table name.
}
\description{
This function retrieves the column names from all tables in a database connection.
It returns a list where each element is a vector of column names, prefixed with the table name.
}
