% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_catalog.R
\name{getConceptCatalogDS}
\alias{getConceptCatalogDS}
\title{Get Concept Catalog}
\usage{
getConceptCatalogDS(resource, tableName)
}
\arguments{
\item{resource}{A resource object representing the database connection.}

\item{tableName}{The name of the table from which to retrieve the concept types.}
}
\value{
A data frame containing the unique concept IDs and their corresponding concept types from the specified table.
}
\description{
This function is called from the DataSHIELD client to fetch a catalog of available concept types from a specific table within
the database. It retrieves the unique concept IDs from the specified table and then fetches the corresponding concept types
from the 'concept' table.
}
