% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_concepts.R
\name{getConceptIdColumns}
\alias{getConceptIdColumns}
\title{Get Concept ID Columns from a Table}
\usage{
getConceptIdColumns(table)
}
\arguments{
\item{table}{A data frame representing the table to be examined.}
}
\value{
A character vector containing the names of the columns that include concept IDs.
}
\description{
This function examines each column in a given table to identify those that contain concept IDs.
Concept ID columns are recognized by the presence of "_concept_id" in their names. These columns
are essential for the translation process, where concept IDs are replaced with their corresponding
concept names.
}
