% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_concepts.R
\name{getConceptIds}
\alias{getConceptIds}
\title{Get Concept IDs from a Table}
\usage{
getConceptIds(table, conceptIdColumns)
}
\arguments{
\item{table}{A data frame representing the table from which to extract concept IDs.}

\item{conceptIdColumns}{A character vector specifying the names of the columns that contain concept IDs.}
}
\value{
A numeric vector containing all unique concept IDs from the specified concept ID columns.
}
\description{
This function extracts all unique concept IDs from specified concept ID columns within a table. It iterates
over each concept ID column, collects all unique concept IDs while omitting any NA values, and returns a
vector of these unique concept IDs. This is crucial for identifying the distinct concepts present in the table
that may require translation or further processing.
}
