% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDictionary.R
\name{getDictionary}
\alias{getDictionary}
\title{Retrieve a dictionary of database tables}
\usage{
getDictionary(connection)
}
\arguments{
\item{connection}{A DBI connection object to the database.}
}
\value{
A list where each element is a data frame representing a table from the database. The names of the list elements are the names of the tables without the "_id" suffix.
}
\description{
This function retrieves a dictionary of ID values based on the IDs identified from the database connection. It iterates over each ID, constructs a SQL query to fetch the corresponding table, and adds it to the dictionary. If a table cannot be fetched, it is ignored.
}
