% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_connection.R
\name{getEmptyColumns}
\alias{getEmptyColumns}
\title{Get Empty Columns}
\usage{
getEmptyColumns(connection, tableName)
}
\arguments{
\item{connection}{A DBI database connection object.}

\item{tableName}{The name of the table to check for empty columns.}
}
\value{
A character vector containing the names of all completely empty columns in the table.
}
\description{
Identifies and retrieves a list of column names from the specified table within the database connection
that are completely empty (i.e., all values are NA or empty strings). This function is useful for data
cleaning and preprocessing steps, allowing users to easily identify and exclude columns that contain no
useful information.
}
