% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getForeignColumnNames.R
\name{getForeignIdColumnNames}
\alias{getForeignIdColumnNames}
\title{Retrieve foreign ID column names from database tables}
\usage{
getForeignIdColumnNames(connection)
}
\arguments{
\item{connection}{A DBI connection object to the database.}
}
\value{
A list of character vectors. Each element of the list corresponds to a table in the database,
and contains the foreign ID column names of that table, excluding 'source' columns.
}
\description{
This function retrieves the names of foreign ID columns from all tables in a database connection.
It returns a list where each element is a vector of foreign ID column names, excluding any columns named 'source'.
}
