% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdCategories.R
\name{getIdCategories}
\alias{getIdCategories}
\title{Get ID categories from database tables}
\usage{
getIdCategories(connection)
}
\arguments{
\item{connection}{A connection object to the database.}
}
\value{
A character vector of ID categories.
}
\description{
This function retrieves the names of all tables in the connected database and appends "_id" to each name.
The resulting list represents the categories of IDs available in the database.
}
