% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_columns.R
\name{getIdColumn}
\alias{getIdColumn}
\title{Get ID Column Name}
\usage{
getIdColumn(tableName)
}
\arguments{
\item{tableName}{The name of the table for which to generate the ID column name.}
}
\value{
A string containing the name of the ID column.
}
\description{
This function generates the name of the ID column based on the provided table name.
It removes the suffix "_occurrence" from the table name, if present, and appends "_id"
to form the ID column name. This is useful for dynamically identifying the ID column
in tables where the base name of the table is used to construct column names.
}
