% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIds.R
\name{getIds}
\alias{getIds}
\title{Retrieve IDs for each category}
\usage{
getIds(connection)
}
\arguments{
\item{connection}{The connection to the database.}
}
\value{
A list of IDs for each category.
}
\description{
This function retrieves the IDs for each category from the 'person' table in the given connection.
It uses the helper function 'getNonRelatedIdColumnNames' to get the column names that are not related to the 'person' table.
Then, it iterates over these column names (categories) and uses the 'getIdsForCategory' function to retrieve the IDs for each category.
The IDs are stored in a list where each element corresponds to a category.
}
