% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsForCategory.R
\name{getIdsForCategory}
\alias{getIdsForCategory}
\title{Retrieve unique IDs for a given category}
\usage{
getIdsForCategory(connection, idCategory)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{idCategory}{A character vector of column names representing an ID category.}
}
\value{
A vector of unique IDs for the given category.
}
\description{
This function retrieves unique IDs for a given category from the database.
It does so by iterating over each column name in the category, fetching the IDs for each column,
and then returning a unique set of these IDs.
}
