% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIdsForColumn.R
\name{getIdsForColumn}
\alias{getIdsForColumn}
\title{Get unique IDs for a given column}
\usage{
getIdsForColumn(connection, columnName)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{columnName}{A string representing the name of the column in the format 'table.column'.}
}
\value{
A list of unique IDs from the specified column.
}
\description{
This function retrieves unique IDs from a specified column in a database table.
The function connects to the database, constructs a SQL query to select distinct IDs from the specified column,
and executes the query. The IDs are then returned as a list.
}
