% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNextRelation.R
\name{getNextRelation}
\alias{getNextRelation}
\title{Get the next relation in the path between two tables}
\usage{
getNextRelation(connection, tableName, targetTableName)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{tableName}{A string representing the name of the starting table.}

\item{targetTableName}{A string representing the name of the target table.}
}
\value{
The name of the next table in the relation path from the starting table to the target table.
}
\description{
This function retrieves the next step in the relations between a given table and a target table.
It is useful for determining the next table to access via IDs when moving towards the target table.
}
