% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonRelatedIdCategories.R
\name{getNonRelatedIdCategories}
\alias{getNonRelatedIdCategories}
\title{Get non-related ID categories from a given table}
\usage{
getNonRelatedIdCategories(connection, tableName)
}
\arguments{
\item{connection}{A connection object to the database.}

\item{tableName}{The name of the table from which to retrieve non-related ID categories.}
}
\value{
A character vector of non-related ID category names.
}
\description{
This function retrieves the categories of non-related IDs from a specified table in the database.
It first identifies the non-related tables by calling the \code{getNonRelatedTables} function, and then appends "_id" to each table name to form the category names.
}
