% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonRelatedIdColumnNames.R
\name{getNonRelatedIdColumnNames}
\alias{getNonRelatedIdColumnNames}
\title{Get non-related ID column names}
\usage{
getNonRelatedIdColumnNames(connection, tableName)
}
\arguments{
\item{connection}{A connection object to the database.}

\item{tableName}{The name of the table for which to retrieve non-related ID column names.}
}
\value{
A character vector of non-related ID column names.
}
\description{
This function retrieves the names of columns that are not related to the specified table.
It does this by first getting all classified ID column names, and then filtering these based on the non-related ID categories.
}
