% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonRelatedTableIdColumnNames.R
\name{getNonRelatedTableIdColumnNames}
\alias{getNonRelatedTableIdColumnNames}
\title{Get non-related table ID column names}
\usage{
getNonRelatedTableIdColumnNames(connection, table, nonRelatedTable)
}
\arguments{
\item{connection}{A connection object to the database.}

\item{table}{A character string specifying the name of the table from which to retrieve ID column names.}

\item{nonRelatedTable}{A character string specifying the name of the non-related table.}
}
\value{
A character vector of non-related ID column names.
}
\description{
This function retrieves the names of ID columns from a specified table that are not related to it.
}
