% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNonRelatedTables.R
\name{getNonRelatedTables}
\alias{getNonRelatedTables}
\title{Retrieve non-related tables}
\usage{
getNonRelatedTables(connection, tableName)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{tableName}{A string representing the name of the table.}
}
\value{
A character vector of table names that are not related to the specified table.
}
\description{
This function retrieves the names of tables that are not related to the specified table in the database.
It first fetches all table names from the database, then identifies the tables related to the specified table.
The function finally returns the names of tables that are neither related to the specified table nor the specified table itself.
}
