% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelatedTables.R
\name{getRelatedTables}
\alias{getRelatedTables}
\title{Retrieve related tables from a database connection.}
\usage{
getRelatedTables(connection, tableName)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{tableName}{The name of the table for which to retrieve related tables.}
}
\value{
A vector of unique table names related to the input table.
}
\description{
This function retrieves the tables related to a given table from a database connection. It does this by first getting the table relations from the database, then it recursively explores the related tables, avoiding cycles by keeping track of already explored tables.
}
