% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRelationPath.R
\name{getRelationPath}
\alias{getRelationPath}
\title{Retrieve the path of relations between two tables}
\usage{
getRelationPath(connection, tableName, targetTableName)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{tableName}{The name of the source table.}

\item{targetTableName}{The name of the target table.}
}
\value{
A vector of table names representing the path of relations from the source table to the target table, or NULL if no such path exists.
}
\description{
This function retrieves the path of relations between a source table and a target table from a database connection. It does this by first getting the table relations from the database, then it recursively explores the related tables, avoiding cycles by keeping track of already explored tables. If the target table is reached, it returns the path of relations. If the target table cannot be reached, it returns NULL.
}
