% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimplifiedConceptsDictionary.R
\name{getSimplifiedConceptsDictionary}
\alias{getSimplifiedConceptsDictionary}
\title{Get a simplified version of the concept dictionary}
\usage{
getSimplifiedConceptsDictionary(connection)
}
\arguments{
\item{connection}{A DBI connection object to the database.}
}
\value{
A list where the concept dictionary is simplified to contain only 'concept_id' and 'concept_name'.
}
\description{
This function retrieves a dictionary from the database and simplifies its concept dictionary by retaining only the 'concept_id' and 'concept_name' columns.
The simplified dictionary is then returned. This function is useful when only the basic information about the concepts is required,
reducing the complexity and size of the dictionary.
}
