% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{Retrieve a table from the database}
\usage{
getTable(connection, tableName)
}
\arguments{
\item{connection}{A DBI connection object to the database.}

\item{tableName}{The name of the table to be retrieved.}
}
\value{
A data frame representing the table, excluding any 'source' columns.
}
\description{
This function retrieves a specified table from the database connection, excluding any columns named 'source'.
}
