% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_catalog.R
\name{getTableCatalogDS}
\alias{getTableCatalogDS}
\title{Get Tables Catalog}
\usage{
getTableCatalogDS(resource)
}
\arguments{
\item{connection}{A DBI database connection object.}
}
\value{
A character vector containing the names of all tables in the database.
}
\description{
This function is called from the DataSHIELD client to fetch a catalog of all available tables within the database.
}
