% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableIdColumnNames.R
\name{getTableIdColumnNames}
\alias{getTableIdColumnNames}
\title{Get ID column names from a table}
\usage{
getTableIdColumnNames(table)
}
\arguments{
\item{table}{A data frame or tibble representing the table from which to extract column names.}
}
\value{
A character vector of column names from the input table that contain '_id'.
}
\description{
This function retrieves the names of columns from a given table that contain '_id' in their names.
These are typically columns that serve as identifiers or keys in a relational database.
The function uses the 'grep' function to match the pattern '_id' in the column names of the table.
}
