% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableNames.R
\name{getTableNames}
\alias{getTableNames}
\title{Get table names}
\usage{
getTableNames(connection)
}
\arguments{
\item{connection}{A connection object to the database.}
}
\value{
A character vector of table names.
}
\description{
This function retrieves the names of all tables in the connected database.
}
