% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTableRelations.R
\name{getTableRelations}
\alias{getTableRelations}
\title{Retrieve table relations from a database connection.}
\usage{
getTableRelations(connection)
}
\arguments{
\item{connection}{A DBI connection object to the database.}
}
\value{
A list of unique column names from each table in the database.
}
\description{
This function retrieves the table relations from a given database connection. It does this by first getting the classified ID column names from the database, then it removes the "_id" suffix from these names. After that, it splits each element of the column names by "." and keeps the first part of the split. Finally, it removes any duplicate names from each list of column names.
}
