% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_connection.R
\name{getTables}
\alias{getTables}
\title{Get Database Tables}
\usage{
getTables(connection)
}
\arguments{
\item{connection}{A DBI database connection object.}
}
\value{
A character vector containing the names of all tables in the database.
}
\description{
Retrieves a list of all tables available in the specified database connection.
This function utilizes the DBI package to list all tables present in the database
connected through the provided DBI connection object.
}
