% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNames.R
\name{makeNames}
\alias{makeNames}
\title{Standardize names for R objects with underscores}
\usage{
makeNames(name)
}
\arguments{
\item{name}{A character string representing the name to be standardized.}
}
\value{
A character string representing the standardized name.
}
\description{
This function takes a name and standardizes it to make it able to be a column name in a data frame.
It converts the name to lower case, replaces periods with underscores, and removes leading and trailing underscores.
The purpose of this function is to ensure that the translated concept names can be reshaped into column names.
}
