% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchIdCategory.R
\name{matchIdCategory}
\alias{matchIdCategory}
\title{Match ID column name to a category}
\usage{
matchIdCategory(idColumnName, idCategories)
}
\arguments{
\item{idColumnName}{A character string representing the name of the ID column.}

\item{idCategories}{A character vector representing the list of categories to match the ID column name against.}
}
\value{
A character string representing the matched category, or NULL if no match is found.
}
\description{
This function matches a given ID column name to a category from a list of categories.
The matching is done based on whether the ID column name ends with the category name.
The function uses the 'grepl' function to match the pattern at the end of the ID column name.
If a match is found, the category name is returned. If no match is found, NULL is returned.
}
