% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTables.R
\name{mergeTables}
\alias{mergeTables}
\title{Merge tables based on a common identifier column}
\usage{
mergeTables(firstTable, secondTable, idColumn)
}
\arguments{
\item{firstTable}{A data frame representing the first table to be merged.}

\item{secondTable}{A data frame representing the second table to be merged.}

\item{idColumn}{A character string representing the name of the common identifier column in both tables.}
}
\value{
A merged data frame.
}
\description{
Merges two tables based on a common identifier column. The function is designed to be used when you have two tables with a common column and you want to merge them into a single table.
}
