% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveColumnNameTokens.R
\name{moveColumnNameTokens}
\alias{moveColumnNameTokens}
\title{Move column name tokens}
\usage{
moveColumnNameTokens(table)
}
\arguments{
\item{table}{A data frame whose column names are to be rearranged.}
}
\value{
A data frame with rearranged column names.
}
\description{
This function rearranges the tokens in the column names of a given table.
The tokens are assumed to be separated by periods. If a column name contains more than one token,
the function moves the first token to the end. The rearranged tokens are then joined back together
with periods to form the new column names.
}
