% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopAchillesDistributionDS}
\alias{omopAchillesDistributionDS}
\title{Get Achilles distribution results (Aggregate)}
\usage{
omopAchillesDistributionDS(omop_symbol, analysis_ids, stratum_filters = NULL)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{analysis_ids}{Integer vector; analysis IDs}

\item{stratum_filters}{Named list; stratum filters}
}
\value{
Data frame with distribution statistics
}
\description{
Get Achilles distribution results (Aggregate)
}
