% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMOPCDMResourceClient.R
\name{OMOPCDMResourceClient}
\alias{OMOPCDMResourceClient}
\title{OMOP CDM resource client}
\description{
A class that provides methods to interact with an OMOP CDM database resource.
It inherits from the ResourceClient class.
}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{OMOPCDMResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPCDMResourceClient-new}{\code{OMOPCDMResourceClient$new()}}
\item \href{#method-OMOPCDMResourceClient-getConnection}{\code{OMOPCDMResourceClient$getConnection()}}
\item \href{#method-OMOPCDMResourceClient-close}{\code{OMOPCDMResourceClient$close()}}
\item \href{#method-OMOPCDMResourceClient-clone}{\code{OMOPCDMResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-new}{}}}
\subsection{Method \code{new()}}{
Initialize the OMOPCDMResourceClient object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$new(resource, dbi.connector = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource to be connected.}

\item{\code{dbi.connector}}{(optional) The DBI connector to be used. If not provided, it will try to find one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the OMOPCDMResourceClient class.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
Get the connection to the resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$getConnection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The connection to the resource.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-close}{}}}
\subsection{Method \code{close()}}{
Close the connection to the resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
