% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopCohortCombineDS}
\alias{omopCohortCombineDS}
\title{Combine cohorts (Assign)}
\usage{
omopCohortCombineDS(omop_symbol, op, cohort_a, cohort_b, new_name = NULL)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{op}{Character; "intersect", "union", or "setdiff"}

\item{cohort_a}{Character; first cohort temp table}

\item{cohort_b}{Character; second cohort temp table}

\item{new_name}{Character; result temp table name}
}
\value{
Character; result temp table name
}
\description{
Combine cohorts (Assign)
}
