% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopCohortGetDefinitionDS}
\alias{omopCohortGetDefinitionDS}
\title{Get a cohort definition (Aggregate)}
\usage{
omopCohortGetDefinitionDS(omop_symbol, cohort_definition_id)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{cohort_definition_id}{Integer}
}
\value{
Named list with definition details
}
\description{
Get a cohort definition (Aggregate)
}
