% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopExpandConceptSetDS}
\alias{omopExpandConceptSetDS}
\title{Expand a concept set (Aggregate)}
\usage{
omopExpandConceptSetDS(omop_symbol, concept_set)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{concept_set}{List; concept set specification}
}
\value{
Integer vector of expanded concept IDs
}
\description{
Expand a concept set (Aggregate)
}
