% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{OMOPResourceClient}
\alias{OMOPResourceClient}
\title{OMOP CDM Resource Client}
\description{
OMOP CDM Resource Client

OMOP CDM Resource Client
}
\details{
R6 class that wraps a DataSHIELD resource pointing to an OMOP CDM database.
Manages the DBI connection and extracts configuration from the resource URL.
}
\keyword{Create}
\keyword{DBI}
\keyword{URL}
\keyword{a}
\keyword{connection}
\keyword{directly}
\keyword{from}
\keyword{internal}
\keyword{parameters}
\keyword{parsed}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{OMOPResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPResourceClient-new}{\code{OMOPResourceClient$new()}}
\item \href{#method-OMOPResourceClient-getConnection}{\code{OMOPResourceClient$getConnection()}}
\item \href{#method-OMOPResourceClient-getParsed}{\code{OMOPResourceClient$getParsed()}}
\item \href{#method-OMOPResourceClient-getDBMS}{\code{OMOPResourceClient$getDBMS()}}
\item \href{#method-OMOPResourceClient-getCDMSchema}{\code{OMOPResourceClient$getCDMSchema()}}
\item \href{#method-OMOPResourceClient-getVocabularySchema}{\code{OMOPResourceClient$getVocabularySchema()}}
\item \href{#method-OMOPResourceClient-getResultsSchema}{\code{OMOPResourceClient$getResultsSchema()}}
\item \href{#method-OMOPResourceClient-getTempSchema}{\code{OMOPResourceClient$getTempSchema()}}
\item \href{#method-OMOPResourceClient-close}{\code{OMOPResourceClient$close()}}
\item \href{#method-OMOPResourceClient-clone}{\code{OMOPResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-new}{}}}
\subsection{Method \code{new()}}{
Create a new OMOP resource client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$new(resource, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A resourcer resource object}

\item{\code{...}}{Reserved for compatibility with the parent resource client.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
Get or create a DBI connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getConnection()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A DBI connection object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getParsed"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getParsed}{}}}
\subsection{Method \code{getParsed()}}{
Get parsed URL parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getParsed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Named list with dbms, host, port, database, schemas
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getDBMS"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getDBMS}{}}}
\subsection{Method \code{getDBMS()}}{
Get the DBMS type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getDBMS()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getCDMSchema"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getCDMSchema}{}}}
\subsection{Method \code{getCDMSchema()}}{
Get the CDM schema name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getCDMSchema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getVocabularySchema"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getVocabularySchema}{}}}
\subsection{Method \code{getVocabularySchema()}}{
Get the vocabulary schema name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getVocabularySchema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getResultsSchema"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getResultsSchema}{}}}
\subsection{Method \code{getResultsSchema()}}{
Get the results schema name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getResultsSchema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-getTempSchema"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-getTempSchema}{}}}
\subsection{Method \code{getTempSchema()}}{
Get the temp schema name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$getTempSchema()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-close}{}}}
\subsection{Method \code{close()}}{
Close the connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
