% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\name{OMOPResourceResolver}
\alias{OMOPResourceResolver}
\title{OMOP CDM Resource Resolver}
\description{
OMOP CDM Resource Resolver

OMOP CDM Resource Resolver
}
\keyword{internal}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{OMOPResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPResourceResolver-isFor}{\code{OMOPResourceResolver$isFor()}}
\item \href{#method-OMOPResourceResolver-newClient}{\code{OMOPResourceResolver$newClient()}}
\item \href{#method-OMOPResourceResolver-clone}{\code{OMOPResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check if this resolver can handle the given resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceResolver$isFor(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A resourcer resource object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Create a new client for the given resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceResolver$newClient(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A resourcer resource object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An OMOPResourceClient, or NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
