% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_reshape.R
\name{rearrangeColumnNames}
\alias{rearrangeColumnNames}
\title{Rearrange Column Name Tokens}
\usage{
rearrangeColumnNames(table)
}
\arguments{
\item{table}{A data frame with column names to be rearranged.}
}
\value{
A data frame with rearranged column names.
}
\description{
This function rearranges the tokens in the column names of a reshaped table. In the reshaped table,
column names consist of tokens separated by ".", where the first token refers to the original table column
and the second token refers to the related concept type. This function modifies the column names so that
the concept type appears first, followed by the original table column name, improving the readability
of the resulting columns.
}
