% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceColumnNames.R
\name{replaceColumnNames}
\alias{replaceColumnNames}
\title{Replace column names in a dictionary using a table column}
\usage{
replaceColumnNames(dictionary, tableColumn, dictionaryColumn)
}
\arguments{
\item{dictionary}{A named list or vector where the names are to be replaced.}

\item{tableColumn}{A character string representing the column in the table that is used to replace the dictionary column names.}

\item{dictionaryColumn}{A character string representing the column in the dictionary whose names are to be replaced.}
}
\value{
A dictionary with replaced column names.
}
\description{
This function replaces the column names in a dictionary using a specific column from a table.
The function identifies the prefix and suffix of the dictionary column in the table column and uses these to construct new column names.
}
