% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshapeTable.R
\name{reshapeTable}
\alias{reshapeTable}
\title{Reshape a table from long to wide format}
\usage{
reshapeTable(table, idColumn, eventColumn)
}
\arguments{
\item{table}{A data frame representing the table to be reshaped.}

\item{idColumn}{A character string representing the id column in the table.}

\item{eventColumn}{A character string representing the event column in the table.}
}
\value{
A data frame representing the reshaped table.
}
\description{
This function reshapes a table from long to wide format using the reShapeDS function from the dsBase package.
The reshaping process is controlled by specifying the id and event columns. The remaining columns are considered as varying variables.
The function also moves the column name tokens to make the reshaped table more understandable.
}
