% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequenceLongitudinalData.R
\name{sequenceLongitudinalData}
\alias{sequenceLongitudinalData}
\title{Sequence longitudinal data}
\usage{
sequenceLongitudinalData(dataTable, idColumn, eventColumn)
}
\arguments{
\item{dataTable}{A data frame representing the longitudinal data to be sequenced.}

\item{idColumn}{A character string representing the id column in the data table.}

\item{eventColumn}{A character string representing the event column in the data table.}
}
\value{
A data frame with sequenced longitudinal data.
}
\description{
This function sequences longitudinal data by numbering repeated variables over time for the same id and event type.
It modifies the event column of the input data table by appending a sequence number to the event name for repeated events.
}
