% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTableCategories.R
\name{sortTableCategories}
\alias{sortTableCategories}
\title{Sorts the columns of a table based on their category.}
\usage{
sortTableCategories(table)
}
\arguments{
\item{table}{A data frame representing the table to be sorted.}
}
\value{
A data frame representing the sorted table.
}
\description{
This function reorders the columns of a table such that columns without a dot in their name come first,
followed by columns with a dot in their name. Within these two categories, columns are further sorted
based on their "_id" suffix. Columns with the same "_id" suffix are grouped together.
}
