% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortTableChronologically.R
\name{sortTableChronologically}
\alias{sortTableChronologically}
\title{Sort a table chronologically}
\usage{
sortTableChronologically(table, timeColumn)
}
\arguments{
\item{table}{A data frame representing the table to be sorted.}

\item{timeColumn}{A character string representing the name of the time column in the table.}
}
\value{
A data frame representing the sorted table.
}
\description{
This function sorts a table based on a specified time column. The table is rearranged in ascending order according to the values in the time column. This is useful when you want to analyze data in a chronological order.
}
