% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_names.R
\name{standardizeColumn}
\alias{standardizeColumn}
\title{Standardize Names in a Column}
\usage{
standardizeColumn(table, column)
}
\arguments{
\item{table}{A data frame representing the table containing the column to be standardized.}

\item{column}{A character string specifying the name of the column whose names are to be standardized.}
}
\value{
A data frame with the specified column's names standardized.
}
\description{
This function applies the \code{standardizeName} function to each name in a specified column of a table.
}
