% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_types.R
\name{stringsToFactors}
\alias{stringsToFactors}
\title{Convert String Columns to Factors in the Table}
\usage{
stringsToFactors(table)
}
\arguments{
\item{table}{A data frame with one or more columns of type character.}
}
\value{
A data frame with character columns converted to factors.
}
\description{
This function takes a data frame and converts all columns of type character to factors.
}
