% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translateColumnConcepts.R
\name{translateColumnConcepts}
\alias{translateColumnConcepts}
\title{Translate column concepts}
\usage{
translateColumnConcepts(connection, table, columnName)
}
\arguments{
\item{connection}{A database connection object.}

\item{table}{A data frame representing the table in which the column concepts are to be translated.}

\item{columnName}{A string representing the name of the column to be translated.}
}
\value{
A data frame with the specified column's concepts translated.
}
\description{
This function translates the concepts in a specified column of a table using a dictionary of simplified concepts.
The dictionary is obtained from a database connection. Each concept in the column is replaced with its corresponding
name from the dictionary. If a concept does not exist in the dictionary or is NA, it is left unchanged.
}
