% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_types.R
\name{conceptsToFactors}
\alias{conceptsToFactors}
\title{Convert Concept ID Columns to Factors in the Table}
\usage{
conceptsToFactors(table)
}
\arguments{
\item{table}{A data frame with one or more columns that contain "concept_id" in their name.}
}
\value{
A data frame with columns that contain "concept_id" in their name converted to factors.
}
\description{
This function takes a data frame and converts all columns that contain "concept_id" in their name to factors.
}
