% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_types.R
\name{convertDateColumns}
\alias{convertDateColumns}
\title{Convert Date and DateTime Columns in the Table}
\usage{
convertDateColumns(table)
}
\arguments{
\item{table}{A data frame with one or more columns that may contain "_date" or "_datetime" in their name.}
}
\value{
A data frame with columns that contain "_datetime" converted to DateTime type and columns that contain "_date"
converted to Date type.
}
\description{
This function takes a data frame and checks each column name by splitting it at each period ('.').
It then converts columns to DateTime type in R if "_datetime" is found in any part of the name except the first token,
and to Date type if "_date" is found under the same conditions, ignoring case.
}
