% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_types.R
\name{convertDateColumns}
\alias{convertDateColumns}
\title{Convert Date Columns to Date Type in the Table}
\usage{
convertDateColumns(table)
}
\arguments{
\item{table}{A data frame with one or more columns that contain "_date" in their name.}
}
\value{
A data frame with columns that contain "_date" in their name converted to Date type.
}
\description{
This function takes a data frame and converts all columns that contain "_date" in their name to Date type in R,
ignoring case.
}
