% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{.achillesAnalysisDomain}
\alias{.achillesAnalysisDomain}
\title{Classify analysis domain from analysis_id range}
\usage{
.achillesAnalysisDomain(analysis_ids)
}
\description{
Classify analysis domain from analysis_id range
}
\keyword{internal}
