% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{.achillesDiscoverIds}
\alias{.achillesDiscoverIds}
\title{Query distinct analysis_ids from achilles_results}
\usage{
.achillesDiscoverIds(handle)
}
\description{
Query distinct analysis_ids from achilles_results
}
\keyword{internal}
