% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achilles.R
\name{.achillesGetDistributions}
\alias{.achillesGetDistributions}
\title{Get Achilles distribution results}
\usage{
.achillesGetDistributions(handle, analysis_ids, stratum_filters = NULL)
}
\arguments{
\item{handle}{CDM handle}

\item{analysis_ids}{Integer vector}

\item{stratum_filters}{Named list; optional stratum filters}
}
\value{
data.frame with analysis_id, stratum_1..5, count_value,
  min/max/avg/stdev/median/p10/p25/p75/p90
}
\description{
Get Achilles distribution results
}
\keyword{internal}
