% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure.R
\name{.assertSafeLevels}
\alias{.assertSafeLevels}
\title{Check if returning distinct levels is safe}
\usage{
.assertSafeLevels(n_levels, n_total)
}
\arguments{
\item{n_levels}{Integer; number of distinct levels}

\item{n_total}{Integer; total number of non-NA values}
}
\value{
TRUE invisibly, or stops with an error
}
\description{
Check if returning distinct levels is safe
}
\keyword{internal}
