% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.classifyConceptRoleHeuristic}
\alias{.classifyConceptRoleHeuristic}
\title{Heuristic concept role classification (no spec)}
\usage{
.classifyConceptRoleHeuristic(table, column)
}
\description{
Heuristic concept role classification (no spec)
}
\keyword{internal}
