% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.coerce_integer64}
\alias{.coerce_integer64}
\title{Execute a SQL query and return a data frame}
\usage{
.coerce_integer64(df)
}
\arguments{
\item{handle}{CDM handle}

\item{sql}{Character; SQL to execute}
}
\value{
Data frame
}
\description{
Execute a SQL query and return a data frame
}
\keyword{internal}
