% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueprint.R
\name{.detectSensitiveColumns}
\alias{.detectSensitiveColumns}
\title{Detect whether a column contains PII / sensitive data}
\usage{
.detectSensitiveColumns(column_name)
}
\arguments{
\item{column_name}{Character; column name}
}
\value{
Logical
}
\description{
Detect whether a column contains PII / sensitive data
}
\keyword{internal}
