% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.extractCohortMembership}
\alias{.extractCohortMembership}
\title{Extract standard OHDSI cohort membership table}
\usage{
.extractCohortMembership(handle, cohort_table, cohort_definition_id)
}
\arguments{
\item{handle}{CDM handle}

\item{cohort_table}{Character; temp table name with cohort members}

\item{cohort_definition_id}{Integer; cohort definition ID}
}
\value{
Data frame with row_id, subject_id, cohort_definition_id,
  cohort_start_date, cohort_end_date
}
\description{
Simple query on the materialized cohort temp table. Produces output
with the standard OHDSI cohort columns.
}
\keyword{internal}
