% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_catalog.R
\name{.omopAnalysisScope}
\alias{.omopAnalysisScope}
\title{Default scope spec for an entry}
\usage{
.omopAnalysisScope(
  accepts_cohort = TRUE,
  accepts_tables = TRUE,
  max_tables = 2L
)
}
\description{
Default scope spec for an entry
}
\keyword{internal}
