% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disclosure.R
\name{.omopDisclosureSettings}
\alias{.omopDisclosureSettings}
\title{Read all disclosure settings from DataSHIELD server options}
\usage{
.omopDisclosureSettings()
}
\value{
Named list of disclosure thresholds and privacy level
}
\description{
Read all disclosure settings from DataSHIELD server options
}
\keyword{internal}
