% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attach hook}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{Library path.}

\item{pkg}{Package name.}
}
\description{
Registers the OMOP CDM resource resolver and displays a startup message
with the package version.
}
\keyword{internal}
