% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{.planExecute}
\alias{.planExecute}
\title{Execute a plan and produce server-side data frames}
\usage{
.planExecute(handle, plan, out_symbols)
}
\arguments{
\item{handle}{CDM handle}

\item{plan}{List; the extraction plan}

\item{out_symbols}{Named list; output name -> R symbol mapping}
}
\value{
Named list of data frames
}
\description{
Execute a plan and produce server-side data frames
}
\keyword{internal}
