% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{.profileColumnStats}
\alias{.profileColumnStats}
\title{Get column-level statistics}
\usage{
.profileColumnStats(handle, table, column)
}
\arguments{
\item{handle}{CDM handle}

\item{table}{Character; table name}

\item{column}{Character; column name}
}
\value{
Named list with column statistics
}
\description{
Get column-level statistics
}
\keyword{internal}
