% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.quoteLiteral}
\alias{.quoteLiteral}
\title{Quote a SQL literal value safely}
\usage{
.quoteLiteral(x)
}
\arguments{
\item{x}{Value to quote}
}
\value{
Character; quoted SQL literal
}
\description{
Quote a SQL literal value safely
}
\keyword{internal}
