% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.toFeatures}
\alias{.toFeatures}
\title{Compute person-level features from event data}
\usage{
.toFeatures(df, table, specs = NULL)
}
\arguments{
\item{df}{Data frame in long format}

\item{table}{Character; source table name}

\item{specs}{Named list of feature specifications}
}
\value{
Data frame with one row per person
}
\description{
Compute person-level features from event data
}
\keyword{internal}
