% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{.toWide}
\alias{.toWide}
\title{Transform a long result to wide format}
\usage{
.toWide(df, table)
}
\arguments{
\item{df}{Data frame in long format}

\item{table}{Character; source table name}
}
\value{
Data frame in wide format
}
\description{
Pivots on the domain concept column (from blueprint).
}
\keyword{internal}
