% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{.vocabExpandConceptSet}
\alias{.vocabExpandConceptSet}
\title{Expand a concept set specification to a full list of concept IDs}
\usage{
.vocabExpandConceptSet(handle, concept_set)
}
\arguments{
\item{handle}{CDM handle}

\item{concept_set}{List with concepts, include_descendants, include_mapped, exclude}
}
\value{
Integer vector of expanded concept IDs
}
\description{
Expand a concept set specification to a full list of concept IDs
}
\keyword{internal}
