% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_schema.R
\name{fillSchemaQuery}
\alias{fillSchemaQuery}
\title{Fill Schema Query}
\usage{
fillSchemaQuery(schema, query)
}
\arguments{
\item{schema}{The name of the schema to be used in the query.}

\item{query}{The schema query template containing the placeholder {schema}.}
}
\value{
The schema query with the {schema} placeholder replaced by the actual schema name.
}
\description{
Fill Schema Query
}
