% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_columns.R
\name{findCaseInsensitiveColumn}
\alias{findCaseInsensitiveColumn}
\title{Find Case-Insensitive Column}
\usage{
findCaseInsensitiveColumn(columnNames, target)
}
\arguments{
\item{columnNames}{A character vector containing the names of the columns to search through.}

\item{target}{A character string representing the name of the column to find, case-insensitively.}
}
\value{
A character string with the name of the target column if found, otherwise NULL.
}
\description{
Given a list of column names, this function searches for a column whose name matches a given target in a case-insensitive manner.
}
