% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table.R
\name{findCaseInsensitiveTable}
\alias{findCaseInsensitiveTable}
\title{Find Case-Insensitive Table}
\usage{
findCaseInsensitiveTable(tableNames, target)
}
\arguments{
\item{tableNames}{A character vector containing the names of the tables to search through.}

\item{target}{A character string representing the name of the table to find, case-insensitively.}
}
\value{
A character string with the name of the target table if found, otherwise NULL.
}
\description{
Given a list of table names, this function searches for a table whose name matches a given target in a case-insensitive manner.
}
