% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_catalog.R
\name{getColumnCatalogDS}
\alias{getColumnCatalogDS}
\title{Get Columns Catalog}
\usage{
getColumnCatalogDS(resource, tableName, dropNA = FALSE)
}
\arguments{
\item{resource}{A resource object representing the database connection.}

\item{tableName}{The name of the table from which to retrieve the column names.}

\item{dropNA}{A logical flag indicating whether to exclude columns that are completely empty (default is FALSE).}
}
\value{
A character vector containing the names of all columns in the specified table, optionally excluding empty ones.
}
\description{
This function is called from the DataSHIELD client to fetch a catalog of all available columns from a specific table within
the database. It allows for an optional filtering to exclude columns that are empty.
}
