% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_connection.R
\name{getColumns}
\alias{getColumns}
\title{Get Table Columns}
\usage{
getColumns(connection, tableName, dropNA = FALSE)
}
\arguments{
\item{connection}{A DBI database connection object.}

\item{tableName}{The name of the table from which to retrieve column names.}

\item{dropNA}{A logical value indicating whether to exclude completely empty columns from the result.}
}
\value{
A character vector containing the names of all columns in the table, optionally excluding empty ones.
}
\description{
Retrieves a list of all column names from the specified table within the database connection.
Optionally, it can exclude columns that are completely empty (NA values) by setting the \code{dropNA} parameter to TRUE.
When \code{dropNA} is TRUE, it first identifies empty columns and then excludes them from the list of returned columns.
}
