% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_columns.R
\name{getConceptIdColumn}
\alias{getConceptIdColumn}
\title{Get Concept ID Column Name}
\usage{
getConceptIdColumn(tableName)
}
\arguments{
\item{tableName}{The name of the table for which to generate the concept ID column name.}
}
\value{
A string containing the name of the concept ID column.
}
\description{
This function generates the name of the concept ID column based on the provided table name.
It removes the suffix "_occurrence" from the table name, if present, and appends "_concept_id"
to form the concept ID column name. This is useful for dynamically identifying the concept ID
column in tables where the base name of the table is used to construct column names.
}
