% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_concepts.R
\name{getConcepts}
\alias{getConcepts}
\title{Get Concept Names from Concept IDs}
\usage{
getConcepts(connection, conceptIds)
}
\arguments{
\item{connection}{A database connection object through which the query will be executed.}

\item{conceptIds}{A numeric vector containing the concept IDs for which names are to be retrieved.}
}
\value{
A data frame with columns "concept_id" and "concept_name", representing the mapping from concept IDs to their names.
}
\description{
This function queries the "concept" table to find the corresponding concept names for a given list of concept IDs.
It constructs a dictionary of concept IDs and their associated names, which is essential for translating concept IDs
in tables to their meaningful names.
}
