% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConnection.R
\name{getConnection}
\alias{getConnection}
\title{Get connection from a resource}
\usage{
getConnection(resource)
}
\arguments{
\item{resource}{An object of class 'OMOPCDMResourceClient' representing the OMOP CDM database resource.}
}
\value{
A connection object to the OMOP CDM database.
}
\description{
This function retrieves the connection from an OMOP CDM database resource.
}
