% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_connection.R
\name{getConnection}
\alias{getConnection}
\title{Get OMOP CDM Database Connection}
\usage{
getConnection(resource)
}
\arguments{
\item{resource}{An object of class OMOPCDMResourceClient representing the OMOP CDM resource client.}
}
\value{
A DBI database connection object.
}
\description{
Retrieves an active database connection from the specified OMOP CDM resource client.
If no connection exists, it creates one using the DBI connector and sets it for future use.
This ensures that database operations are performed over a valid connection.
Additionally, it sets the search path to the schema specified in the resource, if present.
}
