% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table.R
\name{getOMOPCDMTableDS}
\alias{getOMOPCDMTableDS}
\title{Assign OMOP CDM Table to the DataSHIELD Environment}
\usage{
getOMOPCDMTableDS(
  connection,
  tableName,
  conceptFilter = NULL,
  columnFilter = NULL,
  personFilter = NULL,
  mergeColumn = "person_id",
  dropNA = FALSE
)
}
\arguments{
\item{connection}{A DBI database connection object.}

\item{tableName}{The name of the table to be retrieved from the database.}

\item{conceptFilter}{(Optional) A vector of concept IDs to filter the table by specific concepts.}

\item{columnFilter}{(Optional) A vector of column names to keep in the table.}

\item{personFilter}{(Optional) A vector of person IDs to filter the table by specific individuals.}

\item{mergeColumn}{(Optional) The name of the column used for merging tables, defaults to "person_id".}

\item{dropNA}{(Optional) A logical flag indicating whether to drop columns with all NA values, defaults to FALSE.}
}
\value{
A data frame representing the processed table.
}
\description{
This function is designed to be called from the DataSHIELD client. It retrieves a specified table from an OMOP CDM database.
}
\details{
It calls the \code{getTable} function, which is responsible for filtering and performing additional transformation and processing
operations on the table. Once processed, the table is assigned in the DataSHIELD environment.
}
