% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_person.R
\name{getPersonIds}
\alias{getPersonIds}
\title{Get Unique Person IDs from a Table}
\usage{
getPersonIds(table)
}
\arguments{
\item{table}{A data frame or tibble containing the data from which to extract person IDs.}
}
\value{
A vector containing all unique person IDs from the table.
}
\description{
This function extracts all unique person IDs from the specified table. It checks if the table
contains a 'person_id' column and returns a vector of unique person IDs. If the 'person_id'
column is not present, it stops the execution with an error message.
}
