% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_schema.R
\name{getSchemaQueries}
\alias{getSchemaQueries}
\title{Get registered schema queries}
\usage{
getSchemaQueries()
}
\value{
A named list of registered schema queries.
}
\description{
Get registered schema queries
}
