% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_schema.R
\name{getSchemaQuery}
\alias{getSchemaQuery}
\title{Get the schema query for a specific DBMS}
\usage{
getSchemaQuery(dbms)
}
\arguments{
\item{dbms}{The name of the DBMS (e.g., "postgresql", "mysql").}
}
\value{
The schema query corresponding to the specified DBMS.
}
\description{
Get the schema query for a specific DBMS
}
