% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_person.R
\name{getSubsetFilter}
\alias{getSubsetFilter}
\title{Get DataSHIELD's Subset Filter Value}
\usage{
getSubsetFilter()
}
\value{
A numeric value representing the minimum number of patients required for subset operations.
}
\description{
This function retrieves the subset filter value (nfilter.subset) from DataSHIELD's disclosure settings.
The subset filter value determines the minimum number of patients that can result from a subset operation.
It is used as a measure of the minimum number of patients that can be processed, as many tables in OMOP CDM
may contain multiple records related to the same patient (person).
}
