% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_table.R
\name{getTable}
\alias{getTable}
\title{Get a Table from the Database}
\usage{
getTable(
  connection,
  tableName,
  conceptFilter = NULL,
  columnFilter = NULL,
  personFilter = NULL,
  mergeColumn = "person_id",
  dropNA = FALSE,
  sequenceLongitudinal = FALSE
)
}
\arguments{
\item{tableName}{The name of the table to be retrieved, as a character string.}

\item{conceptFilter}{An optional vector of concept IDs for filtering the table.}

\item{columnFilter}{An optional vector specifying which columns to include in the output.}

\item{personFilter}{An optional vector of person IDs for filtering the table.}

\item{mergeColumn}{The column name to be used for merging, defaults to "person_id".}

\item{dropNA}{Whether to exclude columns with all NA values, defaults to FALSE.}

\item{sequenceLongitudinal}{A logical flag indicating whether to sequence longitudinal data, defaults to FALSE.}

\item{resource}{A resource object representing the database connection.}
}
\value{
A data frame containing the filtered table, ready for integration into the DataSHIELD workflow.
}
\description{
This function fetches a specified table from the database for use within the DataSHIELD environment. It allows users to
filter the table based on columns of interest, specific concept types present in the table, and person IDs from another
table in the system. Filtering at the database query level enhances execution times. The function also checks against
DataSHIELD's disclosure control to ensure that the subset operation does not return fewer patients than the minimum allowed.
}
\details{
In cases where the table contains multiple elements with concept types that can be reshaped to a wide format, the function
will automatically reshape the table to facilitate merging with related tables. Users can specify an alternative column
for reshaping (default is "person_id"), useful for tables indirectly related to persons where another column serves as the
identifier in the reshape operation.

Additionally, users have the option to exclude empty tables from the output, further tailoring the dataset to their needs.
}
