% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopAchillesAnalysesDS}
\alias{omopAchillesAnalysesDS}
\title{List Achilles analyses (Aggregate)}
\usage{
omopAchillesAnalysesDS(omop_symbol, domain = NULL)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{domain}{Character; optional domain filter}
}
\value{
Data frame with analysis catalog
}
\description{
List Achilles analyses (Aggregate)
}
