% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopAchillesResultsDS}
\alias{omopAchillesResultsDS}
\title{Get Achilles count results (Aggregate)}
\usage{
omopAchillesResultsDS(
  omop_symbol,
  analysis_ids,
  stratum_filters = NULL,
  min_cell_count = NULL
)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{analysis_ids}{Integer vector; analysis IDs}

\item{stratum_filters}{Named list; stratum filters}

\item{min_cell_count}{Integer; minimum cell count}
}
\value{
Data frame with analysis results
}
\description{
Get Achilles count results (Aggregate)
}
