% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopAchillesStatusDS}
\alias{omopAchillesStatusDS}
\title{Check Achilles availability (Aggregate)}
\usage{
omopAchillesStatusDS(omop_symbol)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}
}
\value{
Named list with availability status
}
\description{
Check Achilles availability (Aggregate)
}
