% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_resource.R
\name{OMOPCDMResourceClient}
\alias{OMOPCDMResourceClient}
\title{OMOP CDM Resource Client}
\description{
This class is designed to manage connections and interactions with OMOP Common Data Model (CDM) resources.
It extends the functionality of the generic ResourceClient class to specifically support OMOP CDM databases.
The class provides methods to initialize the resource client with a specific DBI connector if provided,
otherwise, it attempts to find a suitable connector. It also manages database connections, ensuring
that a connection is established when needed and properly closed when no longer in use.
}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{OMOPCDMResourceClient}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{initialize}}{Constructor method to initialize the OMOPCDMResourceClient object with a specific resource
and an optional DBI connector. If no DBI connector is provided, it attempts to find one. It throws an error
if no suitable DBI connector can be found.}

\item{\code{getConnection}}{Retrieves an active database connection. If no connection exists, it creates one using
the DBI connector and sets it for future use. This ensures that database operations are performed over
a valid connection.}

\item{\code{close}}{Closes the active database connection if it exists and sets the connection to NULL. This method
ensures that resources are properly released when the connection is no longer needed.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{initialize}}{Constructor method to initialize the OMOPCDMResourceClient object with a specific resource
and an optional DBI connector. If no DBI connector is provided, it attempts to find one. It throws an error
if no suitable DBI connector can be found.}

\item{\code{getConnection}}{Retrieves an active database connection. If no connection exists, it creates one using
the DBI connector and sets it for future use. This ensures that database operations are performed over
a valid connection.}

\item{\code{close}}{Closes the active database connection if it exists and sets the connection to NULL. This method
ensures that resources are properly released when the connection is no longer needed.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPCDMResourceClient-new}{\code{OMOPCDMResourceClient$new()}}
\item \href{#method-OMOPCDMResourceClient-getConnection}{\code{OMOPCDMResourceClient$getConnection()}}
\item \href{#method-OMOPCDMResourceClient-close}{\code{OMOPCDMResourceClient$close()}}
\item \href{#method-OMOPCDMResourceClient-clone}{\code{OMOPCDMResourceClient$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asDataFrame"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asDataFrame'><code>resourcer::ResourceClient$asDataFrame()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$new(resource, dbi.connector = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-getConnection"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-getConnection}{}}}
\subsection{Method \code{getConnection()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$getConnection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-close"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
