% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMOPCDMResourceResolver.R
\name{OMOPCDMResourceResolver}
\alias{OMOPCDMResourceResolver}
\title{OMOP CDM resource resolver}
\description{
A class that provides methods to resolve resources in the OMOP Common Data Model (CDM).
It inherits from the ResourceResolver class.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{OMOPCDMResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPCDMResourceResolver-isFor}{\code{OMOPCDMResourceResolver$isFor()}}
\item \href{#method-OMOPCDMResourceResolver-newClient}{\code{OMOPCDMResourceResolver$newClient()}}
\item \href{#method-OMOPCDMResourceResolver-clone}{\code{OMOPCDMResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Checks if the resource is suitable for this resolver.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceResolver$isFor(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource to be checked.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A boolean indicating if the resource is suitable.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Creates a new client for the resource if it is suitable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceResolver$newClient(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{The resource for which a client is to be created.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An instance of the OMOPCDMResourceClient class if the resource is suitable, NULL otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
