% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_resource.R
\name{OMOPCDMResourceResolver}
\alias{OMOPCDMResourceResolver}
\title{OMOP CDM Resource Resolver}
\description{
This class is responsible for resolving resources specific to the OMOP Common Data Model (CDM).
It inherits from the ResourceResolver class and implements methods to determine if a resource
is suitable for OMOP CDM and to create a new client for interacting with OMOP CDM resources.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{OMOPCDMResourceResolver}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{isFor}}{Checks if the given resource is suitable for OMOP CDM based on specific criteria.}

\item{\code{newClient}}{Creates a new OMOPCDMResourceClient for the given resource if it is suitable.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{isFor}}{Checks if the given resource is suitable for OMOP CDM based on specific criteria.}

\item{\code{newClient}}{Creates a new OMOPCDMResourceClient for the given resource if it is suitable.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OMOPCDMResourceResolver-isFor}{\code{OMOPCDMResourceResolver$isFor()}}
\item \href{#method-OMOPCDMResourceResolver-newClient}{\code{OMOPCDMResourceResolver$newClient()}}
\item \href{#method-OMOPCDMResourceResolver-clone}{\code{OMOPCDMResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceResolver$isFor(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceResolver$newClient(resource)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OMOPCDMResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OMOPCDMResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OMOPCDMResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
