% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopCleanupDS}
\alias{omopCleanupDS}
\title{Clean up temp artifacts (Assign)}
\usage{
omopCleanupDS(omop_symbol, prefix = "dsomop_")
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{prefix}{Character; temp table prefix to clean}
}
\value{
Invisible TRUE
}
\description{
Clean up temp artifacts (Assign)
}
