% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopCohortCreateDS}
\alias{omopCohortCreateDS}
\title{Create a cohort (Assign)}
\usage{
omopCohortCreateDS(
  omop_symbol,
  cohort_spec,
  mode = "temporary",
  cohort_id = NULL,
  name = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{cohort_spec}{List; cohort specification DSL}

\item{mode}{Character; "temporary" or "persistent"}

\item{cohort_id}{Integer; cohort definition ID}

\item{name}{Character; cohort name}

\item{overwrite}{Logical}
}
\value{
Character; temp table name or confirmation
}
\description{
Create a cohort (Assign)
}
