% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopCohortListDS}
\alias{omopCohortListDS}
\title{List cohort definitions (Aggregate)}
\usage{
omopCohortListDS(omop_symbol)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}
}
\value{
Data frame with cohort definitions
}
\description{
List cohort definitions (Aggregate)
}
