% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopColumnStatsDS}
\alias{omopColumnStatsDS}
\title{Get column statistics (Aggregate)}
\usage{
omopColumnStatsDS(omop_symbol, table, column)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{table}{Character; table name}

\item{column}{Character; column name}
}
\value{
Named list with column statistics
}
\description{
Get column statistics (Aggregate)
}
