% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopDomainCoverageDS}
\alias{omopDomainCoverageDS}
\title{Get cross-table domain coverage (Aggregate)}
\usage{
omopDomainCoverageDS(omop_symbol)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}
}
\value{
Data frame with per-table person counts
}
\description{
Get cross-table domain coverage (Aggregate)
}
