% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopGetCapabilitiesDS}
\alias{omopGetCapabilitiesDS}
\title{Get capabilities snapshot (Aggregate)}
\usage{
omopGetCapabilitiesDS(omop_symbol)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}
}
\value{
Named list with schema summary and hash
}
\description{
Get capabilities snapshot (Aggregate)
}
