% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopGetDescendantsDS}
\alias{omopGetDescendantsDS}
\title{Get descendant concepts (Aggregate)}
\usage{
omopGetDescendantsDS(omop_symbol, ancestor_ids, include_self = TRUE)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{ancestor_ids}{Numeric vector}

\item{include_self}{Logical}
}
\value{
Data frame with descendant concepts
}
\description{
Get descendant concepts (Aggregate)
}
