% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopListColumnsDS}
\alias{omopListColumnsDS}
\title{List columns (Aggregate)}
\usage{
omopListColumnsDS(omop_symbol, table)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{table}{Character; table name}
}
\value{
Data frame with column metadata
}
\description{
List columns (Aggregate)
}
