% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopListTablesDS}
\alias{omopListTablesDS}
\title{List tables (Aggregate)}
\usage{
omopListTablesDS(omop_symbol)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}
}
\value{
Data frame with table metadata
}
\description{
List tables (Aggregate)
}
