% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopLookupConceptsDS}
\alias{omopLookupConceptsDS}
\title{Lookup concepts by ID (Aggregate)}
\usage{
omopLookupConceptsDS(omop_symbol, concept_ids)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{concept_ids}{Numeric vector}
}
\value{
Data frame with concept details
}
\description{
Lookup concepts by ID (Aggregate)
}
