% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopMissingnessDS}
\alias{omopMissingnessDS}
\title{Get missingness rates (Aggregate)}
\usage{
omopMissingnessDS(omop_symbol, table, columns = NULL)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{table}{Character; table name}

\item{columns}{Character vector; columns to check}
}
\value{
Data frame with missingness rates
}
\description{
Get missingness rates (Aggregate)
}
