% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopPlanPreviewDS}
\alias{omopPlanPreviewDS}
\title{Preview a plan (Aggregate)}
\usage{
omopPlanPreviewDS(omop_symbol, plan)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{plan}{List; the extraction plan}
}
\value{
List with preview information
}
\description{
Preview a plan (Aggregate)
}
