% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopRelationshipGraphDS}
\alias{omopRelationshipGraphDS}
\title{Get relationship graph (Aggregate)}
\usage{
omopRelationshipGraphDS(omop_symbol)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}
}
\value{
Data frame with join relationships
}
\description{
Get relationship graph (Aggregate)
}
