% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopSearchConceptsDS}
\alias{omopSearchConceptsDS}
\title{Search concepts (Aggregate)}
\usage{
omopSearchConceptsDS(
  omop_symbol,
  pattern,
  domain = NULL,
  vocabulary = NULL,
  standard_only = TRUE,
  limit = 50
)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{pattern}{Character; search pattern}

\item{domain}{Character; domain filter}

\item{vocabulary}{Character; vocabulary filter}

\item{standard_only}{Logical; only standard concepts}

\item{limit}{Integer; max results}
}
\value{
Data frame with concept results
}
\description{
Search concepts (Aggregate)
}
