% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopTableStatsDS}
\alias{omopTableStatsDS}
\title{Get table statistics (Aggregate)}
\usage{
omopTableStatsDS(omop_symbol, table, stats = c("rows", "persons"))
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{table}{Character; table name}

\item{stats}{Character vector; which stats to return}
}
\value{
Named list with safe statistics
}
\description{
Get table statistics (Aggregate)
}
