% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{omopValueCountsDS}
\alias{omopValueCountsDS}
\title{Get value counts (Aggregate)}
\usage{
omopValueCountsDS(
  omop_symbol,
  table,
  column,
  top_n = 20,
  suppress_small = TRUE
)
}
\arguments{
\item{omop_symbol}{Character; the OMOP handle symbol}

\item{table}{Character; table name}

\item{column}{Character; column name}

\item{top_n}{Integer}

\item{suppress_small}{Logical}
}
\value{
Data frame with value counts
}
\description{
Get value counts (Aggregate)
}
