% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_schema.R
\name{registerSchemaQueries}
\alias{registerSchemaQueries}
\title{Register multiple schema queries}
\usage{
registerSchemaQueries(queries, override = TRUE)
}
\arguments{
\item{queries}{A named list of schema queries where names are DBMS names and values are queries.}

\item{override}{A logical value indicating whether to override existing queries for the DBMS.}
}
\description{
Register multiple schema queries
}
