% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_reshape.R
\name{reshapeTable}
\alias{reshapeTable}
\title{Reshape Table to Wide Format}
\usage{
reshapeTable(table, conceptIdColumn, mergeColumn)
}
\arguments{
\item{table}{A data frame representing the table to be reshaped.}

\item{conceptIdColumn}{A string specifying the name of the concept ID column, which characterizes the types of elements.}

\item{mergeColumn}{A string specifying the name of the merge column, typically "person_id", used as the identifier variable.}
}
\value{
A data frame in wide format with standardized and sequenced column names based on the concept ID column.
}
\description{
This function reshapes a given table to a wide format, which facilitates merging with the rest of the tables in the
database using the specified 'merge column'.
}
\details{
It standardizes the names in the concept ID column since these names will become column names in the reshaped table and
must be compatible with the data frame column naming system.

It also sequences repeated elements of the same type for the same person in the columns to separate these instances
for the reshaping process. This is necessary to handle repeated concept names and longitudinal data.

The reshaping process uses the merge column as the identifier variable (idvar), which is typically "person_id" unless
another connecting element to the rest of the database is specified. The time variable (timevar) is the concept ID column,
characterizing the types of elements for each row in the table. This creates a merge column x concept ID column relationship.
}
