% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_longitudinal.R
\name{sequenceColumn}
\alias{sequenceColumn}
\title{Sequence Column in a Table}
\usage{
sequenceColumn(table, conceptIdColumn, mergeColumn)
}
\arguments{
\item{table}{A data frame representing the table to be processed.}

\item{conceptIdColumn}{A character string specifying the name of the column that contains concept IDs.}

\item{mergeColumn}{A character string specifying the name of the column used for merging.}
}
\value{
A data frame with sequence numbers appended to the concept ID column for rows that have the same
combination of concept ID column and merge column values.
}
\description{
This function adds a sequence number to rows within a table that have the same combination of concept ID column
and merge column values. It is particularly useful for distinguishing between multiple occurrences of the same
entity in longitudinal data. The sequence number is appended to the concept ID column for duplicates, starting
from the second occurrence.
}
