% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_names.R
\name{standardizeName}
\alias{standardizeName}
\title{Standardize Name}
\usage{
standardizeName(name)
}
\arguments{
\item{name}{A character string representing the name to be standardized.}
}
\value{
A character string representing the standardized name.
}
\description{
This function standardizes a given name to ensure compatibility with the translation process.
It performs the following operations:
\enumerate{
\item Converts the name to a valid R object name (make.names)
\item Changes all characters to lowercase
\item Replaces periods with underscores
\item Removes repeated underscores
\item Trims leading and trailing underscores
}
}
