% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_concepts.R
\name{translateConcepts}
\alias{translateConcepts}
\title{Replace Concept IDs with Concept Names in a Table}
\usage{
translateConcepts(table, conceptIdColumns, concepts)
}
\arguments{
\item{table}{A data frame representing the table in which concept IDs are to be translated.}

\item{conceptIdColumns}{A character vector specifying the names of the columns that contain concept IDs.}

\item{concepts}{A data frame serving as a dictionary, mapping concept IDs to concept names, obtained from the "concept" table.}
}
\value{
A data frame with concept IDs replaced by concept names in the specified columns.
}
\description{
This function replaces concept IDs with their corresponding concept names within a table. For each column that contains concept IDs,
it substitutes the concept ID values with the concept names using a provided dictionary (concepts) obtained from the "concept" table
in the database. This is essential for converting concept IDs into a more human-readable format.
}
