% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_concepts.R
\name{translateTable}
\alias{translateTable}
\title{Translate Table Concepts}
\usage{
translateTable(connection, table)
}
\arguments{
\item{connection}{A DBI database connection object.}

\item{table}{A data frame representing the table with concept IDs to be translated.}
}
\value{
A data frame with concept IDs replaced by their corresponding concept names.
}
\description{
This function translates concept IDs in a table to their corresponding concept names.
}
\details{
It first checks if the 'concept' table exists in the database. If it does, the function identifies columns containing
concept IDs, retrieves these IDs, and then fetches the corresponding concept names from the 'concept' table.

Finally, it replaces the concept IDs in the original table with the fetched concept names.
}
